requirements.txt
Flask[async]==2.3.3
Flask-SQLAlchemy==3.1.1
python-docx==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
edge-tts==6.1.9
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
